Changed cg_mshift,MINRES, MINRES_mshift to have the same behavior as g5QMR_mshift: now all these inverters explicitly check the solutions and call the non-multishift version of itself on those vector which have not reach convergence. All vectors returned are now garanteed to have the requested precision. If the required precision is too high these functions never exit.
Logger: now it is posible to assign different verbosity level to IDs.

Quark propagators: now the functions have an additional parameter: the desired accuracy.

Added logger calls in the code.



Detailed list of changes follow:

* Added some log in find_spec_limits.c,MD integrators,inverters,rational approx,update rhmc,force routines,qspectrum.
* Added matrix-vector multiplication counter in Dphi.c
* Modified cg_mshift.c and MINRES_mshift.c to check each solution and use non-multishift as a fallback.
* Modified MINRES.c to call itself until required precision is obtained.
* Corrected a bug in check6.c and check7.c due to the new inverter implementation.
* Modified eva.c to use logger. Changed check9.c and check10.c to enable logging.
* error.c now uses logger.
* rhmc.c now use logger.
* logger.c now has id specific levels;
* quark_propagator.c changed so that user can easily specify the propagator accuracy.



git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/HiRep@27 37a69121-b030-0410-a54a-fdca83ffee47

+++ Include/update.h
@@ -18,7 +18,7 @@ void Force0(float dt, suNg_algebra_vector *force);
 void Force(float dt, suNg_algebra_vector *force);
 
 typedef struct {
-	float tlen; /* trajectory lenght */
+	double tlen; /* trajectory lenght */
 	unsigned int nsteps; /* number of step in the integration */
 	unsigned int gsteps; /* number of substeps for the gauge part every step */
 } int_par;
@@ -31,9 +31,9 @@ void gaussian_spinor_field(suNf_spinor *s);
 
 typedef struct {
   /* sim parameters */
-  float beta;
+  double beta;
   int nf;
-  float mass;
+  double mass;
 	
 	double MT_prec; /* metropolis test precision */
 	double MD_prec; /* molecular dynamics precision */
@@ -46,7 +46,16 @@ typedef struct {
 } rhmc_par;
 void init_rhmc(rhmc_par *par);
 void free_rhmc();
+
+/* update the gauge field using RHMC algorithm
+ * return code: (<0 means an error has occurred)
+ * 0 => conf correctly generated but has not passed metropolis test
+ * 1 => conf accepted
+ * 
+ * -1 => rhmc has not been initialized. call init_rhmc first.
+ */
 int update_rhmc();
+
 /* this is the basic operator used in the update */
 void H2(suNf_spinor *out, suNf_spinor *in);
 void Force_rhmc_f(float dt, suNg_algebra_vector *force);
